estudando operadores aritméticos

+++ pythonexercicios/ex007.py
@@ -1,5 +1,5 @@
-n1=int(input('primeira avaliação: '))
-n2=int(input('segunda avaliação: '))
+n1=float(input('primeira avaliação: '))
+n2=float(input('segunda avaliação: '))
 s=n1+n2
 m=s/2
 print(f'A média é {m}')
